pyproject.toml
[tool.poetry]
name = "render"
version = "0.3.0"
description = "render thing I did for my bot"
authors = ["applejuice <30875598+applejuice-wastaken@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
imageio = "^2.15.0"
Pillow = "^9.0.1"
numpy = "^1.22.2"
pytweening = "^1.0.4"
pyopencl = "^2022.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]

opencl = [
    "pyopencl"
]